+++ db/migrate/20130507184659_remove_email_fields_from_designs.rb
@@ -1,0 +1,11 @@
+class RemoveEmailFieldsFromDesigns < ActiveRecord::Migration
+  def up
+    remove_column :designs, :email_subject_template
+    remove_column :designs, :email_template
+  end
+
+  def down
+    add_column :designs, :email_template, :text
+    add_column :designs, :email_subject_template, :string
+  end
+end