+++ db/migrate/20130507191216_remove_emails_from_sites.rb
@@ -1,0 +1,9 @@
+class RemoveEmailsFromSites < ActiveRecord::Migration
+  def up
+    remove_column :sites, :emails
+  end
+
+  def down
+    add_column :sites, :emails, :text
+  end
+end