+++ test/helpers/sheet_emails_helper_test.rb
@@ -1,4 +1,0 @@
-require 'test_helper'
-
-class SheetEmailsHelperTest < ActionView::TestCase
-end