+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   - Updated to rails 4.0.0.rc1
   - Updated to contour 2.0.0.beta.8
 
+### Refactoring
+- Removed SheetEmails in favor of Sheet Comments
+- Removed Batch Sheets in favor of Design Surveys
+- Removed Site Emails in favor of Site Members who are notified via sheet comments
+
 ### Bug Fix
 - Fixed Recent Entries with long subject codes not being properly hidden in Firefox